Getting some changes from alembicmaya-experimental, it contains:
Some #include order changes, and switching from size_t to unsigned int to remove
some Windows warnings under certain setups.

Make argument parsing less fragile in AbcExport.

Use Maya's workspaces to resolve file names when appropriate.

+++ maya/AbcImport/util.cpp
@@ -232,7 +232,8 @@ bool getDagPathByChildName(MDagPath & ioDagPath, const std::string & iChildName)
             return true;
         }
 
-        unsigned int endLength = name.length() - iChildName.length();
+        unsigned int endLength = (unsigned int)(name.length() - 
+            iChildName.length());
         if ((name.length() > iChildName.length() + 1) && 
             (name[endLength - 1] == '|' || name[endLength - 1] == ':') &&
             (name.substr(endLength) == iChildName))
